+++ tosho/src/impl/sjv/download.rs
@@ -346,9 +346,13 @@ pub(crate) async fn sjv_download(
                 // create chapter dir
                 std::fs::create_dir_all(&image_dir).unwrap();
 
-                let total_image_count = chapter.pages as u64;
+                // Determine total image count, if we start at 0
+                // then the total image count is the same as the chapter.pages
+                // If above 0, then we need to add that amount to the total image count
+                let start_page = chapter.start_page;
+                let total_image_count = chapter.pages + start_page;
 
-                let progress = Arc::new(indicatif::ProgressBar::new(total_image_count + 1));
+                let progress = Arc::new(indicatif::ProgressBar::new(total_image_count.into()));
                 progress.enable_steady_tick(std::time::Duration::from_millis(120));
                 progress.set_style(
                     indicatif::ProgressStyle::with_template(
@@ -360,9 +364,7 @@ pub(crate) async fn sjv_download(
                 );
                 progress.set_message("Downloading");
 
-                let max_page = chapter.pages + 1;
-
-                let tasks: Vec<_> = (0..max_page)
+                let tasks: Vec<_> = (0..total_image_count)
                     .map(|page| {
                         // wrap function in async block
                         let wrap_client = client.clone();
@@ -370,7 +372,6 @@ pub(crate) async fn sjv_download(
                         let cnsl = console.clone();
                         let progress = Arc::clone(&progress);
                         let chapter_id = chapter.id;
-                        let max_page = max_page;
                         tokio::spawn(async move {
                             let download_url = wrap_client
                                 .get_manga_url(chapter_id, false, Some(page))
@@ -395,14 +396,7 @@ pub(crate) async fn sjv_download(
                             match wrap_client.stream_download(&download_url, writer).await {
                                 Ok(_) => {}
                                 Err(err) => {
-                                    // If page is not the last page, then we should error out
-                                    // Since sometimes the last page is not available?
-                                    if page < max_page {
-                                        cnsl.error(&format!(
-                                            "    Failed to download image: {}",
-                                            err
-                                        ));
-                                    }
+                                    cnsl.error(&format!("    Failed to download image: {}", err));
                                     // silent delete the file
                                     tokio::fs::remove_file(&img_dl_path)
                                         .await